+++ src/pages/onboarding/FeaturedBrand.tsx
@@ -1,7 +1,8 @@
 import React from "react"
 import { Swiper, SwiperSlide } from "swiper/react"
 import { Box, Typography } from "@mui/material"
-import { NaviWrap, PrimaryButton } from "../../components/navigationbar.styles"
+import { NaviWrap } from "../../components/navigationbar.styles"
+import { PrimaryButton } from "../../styles/buttons.styles"
 import { SwiperWrap } from "./featuredBrand.styles"
 import { useNavigate } from "react-router-dom"
 import "swiper/css"
