+++ src/pages/shop/productView.styles.ts
@@ -1,6 +1,7 @@
 import { css } from "@emotion/css"
 import { styled } from "@mui/material/styles"
-import { Box } from "@mui/material"
+import { Box, Button } from "@mui/material"
+import { PrimaryButton } from "../../styles/buttons.styles"
 
 export const ProductMainInfo = styled(Box)(() => ({
   '& > .product_brand': {
@@ -32,8 +33,24 @@ export const ProductMainInfo = styled(Box)(() => ({
   '& > .product_delivery': {
     display: 'flex',
     justifyContent: 'space-between',
-    '& .button': {
-      fontSize: '.875rem',
+    '& > .MuiBox-root': {
+      flex: '1 1 50%',
     },
+
   }
+}))
+
+export const BrandInfo = styled(Button)(() => ({
+  display: 'flex',
+  alignItems: 'center',
+  padding: '8px 0',
+  border: '1px solid #DADADA',
+  borderRadius: 0,
+  textAlign: 'left'
+}))
+
+export const OrderButton = styled(PrimaryButton)(() => ({
+  padding: '.825rem .5rem',
+  fontSize: '.875rem',
+  borderRadius: 0,
 }))
